fix memmory leak

+++ src/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: edrodrig <edrodrig@student.42sp.org.b      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/01 01:27:06 by edrodrig          #+#    #+#             */
-/*   Updated: 2022/05/01 01:31:06 by edrodrig         ###   ########.fr       */
+/*   Updated: 2022/05/01 22:22:43 by edrodrig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int ft_strlen(const char *str)
+size_t ft_strlen(const char *str)
 {
 	int count;
